requirements.txt
alembic==1.4.3
aniso8601==8.0.0
astroid==2.4.2
attrs==20.2.0
Authlib==0.15.2
autopep8==1.5.4
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
Flask==1.1.2
Flask-Cors==3.0.9
flask-marshmallow==0.14.0
Flask-Migrate==2.5.3
flask-redis==0.4.0
flask-restx==0.2.0
Flask-SQLAlchemy==2.4.4
idna==2.10
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.8.0
marshmallow-sqlalchemy==0.24.0
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pycparser==2.20
pylint==2.6.0
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.6
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.15.0
python-editor==1.0.4
pytz==2020.1
redis==3.5.3
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.20
toml==0.10.1
urllib3==1.25.11
Werkzeug==1.0.1
wrapt==1.12.1
